Move module, seal ArrayType, docs, unreachables
